requirements.txt
# ==============================================
# Dependencias del proyecto Farkim
# Instalar con: pip install -r requirements.txt
# ==============================================

# Dashboard web
streamlit==1.55.0
streamlit-authenticator==0.4.2

# Gráficos interactivos
plotly==6.6.0

# Procesamiento de datos
pandas==2.3.3
openpyxl==3.1.5
xlrd==2.0.2

# Google Sheets API
gspread==6.2.1
google-auth==2.49.1
google-auth-oauthlib==1.3.0

# Seguridad y autenticación
bcrypt==5.0.0
PyYAML==6.0.3

# HTTP
requests==2.32.5
